test(op-reth): move txpool maintenance tests
